ghci: add fedora mirror container
This adds a new dockerfile called `fedmir`. It takes as input a set of
fedora packages, an architecture, as well as a fedora release version,
and then downloads the specified packages and creates a fedora
repository. It then uses the official `httpd` image as base and serves
this manually created fedora image.

This container allows to create custom fedora mirrors with a selected
set of packages. The idea is to build these with all packages needed
for our tests, pushing it to `Github Packages`, and then using it in our
`Github Actions` infrastructure to avoid accessing resources from
outside the CI system.