actions/ghci-osbuild: expose host to container
Expose the entire host file-system to the container as `/ci/host`. This
is mainly meant for easy debugging and bootstrapping of new features.
This way, combined with `--privileged` and `--net=host`, we can do
basically anything in the container.

Note that we do not use this for sandboxing, so there are no security
implications here.