containers/koji: add Kerberos/GSSAPI auth support
Previously, users were only able to authenticate using username & password.
This does not reflect the production instances of koji though because as far
as I know, none of them supports this particular authentication type.

This commit adds support for Kerberos/GSSAPI/SPNEGO authentication type. Note
that you need a working KDC for this setup to work. See the README in
Dockerfile.