Add new "cloud-tools" container image.
With RHEL-9, the osbuild-composer run into issues related to the CLI
tools provided by each cloud provider. In general they can not be
installed on the RHEL-9 host. This commit adds a new Fedora-based image
with all the necessary CLI tools (for AWS, Azure and GCP), which can be
used in osbuild-composer CI tests to interact with the cloud-provider's
API.

Signed-off-by: Tomas Hozza <thozza@redhat.com>