images/osbuild-ci: disable git-safe-dir
Git recently stopped supporting running `git` from within a
directory-path that contains `.git/` ancestors with a different UID.
This is to prevent malicious users from injecting code into git-hooks
of a different user.

We disable this security measure for the CI, since we do not invoke git
from untrusted paths.

Signed-off-by: David Rheinsberg <david.rheinsberg@gmail.com>