Add osbuild-ci images based on c8s and c9s
Add two new variants of the osbuild-ci image based on c8s and c9s.

CentOS Stream is lacking some packages. Install all of the unavailable
python packages using pip and also pytest on cXs images. The reason is
for not installing pytest from cXs repositories is to get the latest
version of it to not have to deal with lack of features on older
releases.

In addition, the cXs base images contain minimal variants of some
packages, which we want to install into them. To resolve the conflicts,
allow erasing of packages during package installation.

The intention is that these cXs images will be used mostly to run the
unit tests. The Fedora image should be used for all the additional
testing (such as pylint, mypy, isort, ...), including unit tests.

Signed-off-by: Tomáš Hozza <thozza@redhat.com>