src/scripts/dnf.sh: allow erasing packages on installation
Some base images contain minimal variants of packages, which conflict
with the regular packages that we want to install into our image. The
only sensible way to do that is to allow erasing of packages when
installing other packages.

Add an optional script argument to allow erasing packages on
installation.

Signed-off-by: Tomáš Hozza <thozza@redhat.com>