Use 'osbuild-ci' name for the Fedora image
When I was adding the c8s and c9s variants of the osbuild-ci image, I
first added the '-fedora' suffix to the 'osbuild-ci' image which is
based on Fedora. However, I later reverted it back, basically to keep
the current version of osbuild CI working without any change.
Unfortunately, it turns out that I missed to revert one case of this
change. Fixing it now.

Signed-off-by: Tomáš Hozza <thozza@redhat.com>