Google Cloud SDK repo: use el9 and don't force architecture
Something must have happened to the GCP repository, because we have been
building the 'cloud-tools' container for both aarch64 and x86_64 for the
past 2 years, but using a GCP repository that hard-coded the x86_64 arch
in its URL. And so far everything worked just fine. However, the aarch64
builds now started to fail on installation of x86_64 packages.

Use the el0 version of the repository and don't specify the architecture
explicitly in the URL, but let DNF substitute it based on the host arch.

Signed-off-by: Tomáš Hozza <thozza@redhat.com>