src/scripts/dnf.py: allow using --nobest and use it for osbuild-ci-c10s
The c10s container contains newer packages than the ones that are
available in the repositories. Therefore we are running into a situation
when DNF would need to downgrade some packages, in order to install the
requested package set. This fails, unless the --nobest options is passed
to DNF.

Extend the dnf.py script and osbuild-ci.Dockerfile, to allow using the
--nobest option if needed. Use the option for the c10s osbuild-ci
container.

Signed-off-by: Tomáš Hozza <thozza@redhat.com>