GHA: rebuild and deploy all images on schedule twice a month
We often forget to rebuild the container images. Make the rebuild
automatic (twice a month). This will allow us to then also trigger
workflows in repositories that use these container images. That should
hopefully help us to keep testing with the latest software and not have
to think about this.

Signed-off-by: Tomáš Hozza <thozza@redhat.com>